Feat: mysql까지 sink 완료
