--- aiProcessor/src/main/java/org/newshabit/app/aiProcessor/application/port/RefineNewsInputPort.java
@@ -1,8 +1,0 @@
-package org.newshabit.app.aiProcessor.application.port;
-
-import org.newshabit.app.aiProcessor.domain.RefinedNews;
-import org.newshabit.app.common.domain.entity.CrawledNews;
-
-public interface RefineNewsInputPort {
-	RefinedNews refineCrawledNews(CrawledNews crawledNews);
-}
