--- aiProcessor/src/main/java/org/newshabit/app/aiProcessor/application/port/RefinedNewsRepositoryOutputPort.java
@@ -1,5 +1,0 @@
-package org.newshabit.app.aiProcessor.application.port;
-
-public interface RefinedNewsRepositoryOutputPort {
-	boolean existsByOriginalUrl(String url);
-}
