+++ common/build.gradle
@@ -8,6 +8,8 @@ jar { enabled = true }
 dependencies {
     implementation 'org.springframework.cloud:spring-cloud-stream'
     implementation 'org.springframework.cloud:spring-cloud-stream-binder-kafka'
+    compileOnly 'org.springframework.boot:spring-boot-starter-data-jpa'
+
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.cloud:spring-cloud-stream-test-binder'
