--- common/src/main/java/org/newshabit/app/common/application/port/TestOutputPort.java
@@ -1,6 +1,0 @@
-package org.newshabit.app.common.application.port;
-
-public interface TestOutputPort {
-
-    public void doSomething();
-}
