--- common/src/main/java/org/newshabit/app/common/infrastructure/adapter/TestAdapter.java
@@ -1,13 +1,0 @@
-package org.newshabit.app.common.infrastructure.adapter;
-
-import org.newshabit.app.common.application.port.TestOutputPort;
-import org.springframework.stereotype.Component;
-
-
-@Component
-public class TestAdapter implements TestOutputPort {
-    @Override
-    public void doSomething() {
-        System.out.println("Test Adapter");
-    }
-}
