+++ crawl/src/main/java/org/newshabit/app/crawl/application/port/MessageInputPort.java
@@ -1,7 +1,7 @@
 package org.newshabit.app.crawl.application.port;
 
 import java.util.List;
-import org.newshabit.app.common.domain.entity.CrawledNews;
+import org.newshabit.app.common.domain.model.CrawledNews;
 
 public interface MessageInputPort {
 	void publishCrawledNews(List<CrawledNews> newsList);
