+++ kafka/scripts/init-kafka.sh
@@ -8,7 +8,7 @@ echo "init kafka start"
   --create --topic crawl --partitions 3 --replication-factor 3
 
 /bin/kafka-topics --bootstrap-server broker-1:19092,broker-2:19092,broker-3:19092 \
-  --create --topic refined-news --partitions 3 --replication-factor 3
+  --create --topic refinedNews --partitions 3 --replication-factor 3
 
 sleep 1
