+++ settings.gradle
@@ -2,5 +2,5 @@ rootProject.name = 'app'
 include 'crawl'
 include 'news'
 include 'common'
-include 'aiProcessor'
+include 'pubsub'
